# Streaming PromQL engine benchmark tooling