--- a/pkg/streamingpromql/benchmarks/comparison_test.go
+++ b/pkg/streamingpromql/benchmarks/comparison_test.go
@@ -89,7 +89,7 @@ func BenchmarkQuery(b *testing.B) {
 	}
 }
 
-func TestBenchmarkQueries(t *testing.T) {
+func TestBothEnginesReturnSameResultsForBenchmarkQueries(t *testing.T) {
 	metricSizes := []int{1, 100} // Don't bother with 2000 series test here: these test cases take a while and they're most interesting as benchmarks, not correctness tests.
 	q := createBenchmarkQueryable(t, metricSizes)
 	cases := TestCases(metricSizes)