#### What this PR does

This PR adds a tool to capture peak memory utilisation of the PromQL engine benchmarks added in #7693.

It also refactors the benchmarks to use a Mimir ingester rather than a TSDB directly. This makes it possible to run the benchmarks with the ingester out-of-process. This is more representative of how the query engine will be used in the real world.

The `benchmark-query-engine` tool takes advantage of this to run the ingester in its process, and run the benchmark in a separate process, ensuring that the peak memory utilisation recorded by `benchmark-query-engine` does not include memory used by the ingester. It captures the peak RSS of the benchmark process and reports this as the peak memory utilisation of the benchmark. There's more details about `benchmark-query-engine` in [the readme](https://github.com/grafana/mimir/blob/charleskorn/peak-memory-benchmarking-take-3/tools/benchmark-query-engine/README.md).

It produces output like the following, allowing it to be used with tools like `benchstat`, as well as [the engine benchmark comparison script](https://github.com/grafana/mimir/blob/main/pkg/streamingpromql/compare.sh):

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/streamingpromql/benchmarks
BenchmarkQuery/a_1,_instant_query/streaming-10         	   10228	    115452 ns/op	   17640 B/op	     302 allocs/op     65372160 B
BenchmarkQuery/a_1,_instant_query/standard-10         	    8766	    122042 ns/op	   20730 B/op	     362 allocs/op     66224128 B
```

<details>
<summary>Full benchmark results</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/streamingpromql/benchmarks
                                                                 │      standard      │              streaming              │
                                                                 │       sec/op       │    sec/op     vs base               │
Query/a_1,_instant_query                                                 123.9µ ± 14%   115.3µ ±  2%   -6.93% (p=0.002 n=6)
Query/a_1,_range_query_with_1_step                                       125.9µ ±  8%   113.4µ ±  1%   -9.94% (p=0.002 n=6)
Query/a_1,_range_query_with_100_steps                                    130.4µ ±  2%   118.6µ ±  2%   -9.02% (p=0.002 n=6)
Query/a_1,_range_query_with_1000_steps                                   187.6µ ±  2%   178.7µ ±  3%   -4.76% (p=0.002 n=6)
Query/a_100,_instant_query                                               612.6µ ±  4%   572.2µ ±  1%   -6.61% (p=0.002 n=6)
Query/a_100,_range_query_with_1_step                                     608.4µ ±  4%   577.6µ ±  4%   -5.06% (p=0.009 n=6)
Query/a_100,_range_query_with_100_steps                                  1.093m ±  3%   1.028m ±  1%   -5.93% (p=0.002 n=6)
Query/a_100,_range_query_with_1000_steps                                 5.573m ±  4%   5.091m ±  1%   -8.65% (p=0.002 n=6)
Query/a_2000,_instant_query                                              7.665m ±  5%   7.178m ±  3%   -6.35% (p=0.009 n=6)
Query/a_2000,_range_query_with_1_step                                    7.883m ±  7%   7.300m ±  1%   -7.39% (p=0.002 n=6)
Query/a_2000,_range_query_with_100_steps                                 15.89m ±  7%   15.06m ±  1%   -5.23% (p=0.002 n=6)
Query/a_2000,_range_query_with_1000_steps                                97.36m ±  6%   88.61m ±  1%   -8.99% (p=0.002 n=6)
Query/rate(a_1[1m]),_instant_query                                       128.4µ ±  3%   120.5µ ± 12%        ~ (p=0.132 n=6)
Query/rate(a_1[1m]),_range_query_with_1_step                             132.2µ ±  5%   119.0µ ±  7%   -9.99% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_100_steps                          139.8µ ±  4%   124.7µ ±  9%  -10.84% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_1000_steps                         256.9µ ± 10%   185.1µ ±  2%  -27.93% (p=0.002 n=6)
Query/rate(a_100[1m]),_instant_query                                     594.7µ ±  2%   568.5µ ±  3%   -4.41% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_1_step                           602.8µ ±  4%   568.9µ ±  1%   -5.62% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_100_steps                        1.630m ±  3%   1.175m ± 11%  -27.91% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_1000_steps                      10.325m ±  1%   5.856m ±  1%  -43.28% (p=0.002 n=6)
Query/rate(a_2000[1m]),_instant_query                                    7.321m ±  4%   6.915m ±  1%   -5.55% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_1_step                          7.455m ±  3%   7.576m ±  5%        ~ (p=0.699 n=6)
Query/rate(a_2000[1m]),_range_query_with_100_steps                       26.51m ±  2%   17.76m ±  0%  -33.01% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_1000_steps                      194.3m ±  6%   104.4m ±  1%  -46.25% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_10000_steps                       1190.5µ ± 11%   752.1µ ±  0%  -36.83% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_10000_steps                      96.20m ±  1%   53.19m ±  1%  -44.71% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_10000_steps                      1.974 ±  1%    1.067 ±  6%  -45.93% (p=0.002 n=6)
Query/rate(a_1[1d]),_instant_query                                       771.6µ ± 15%   543.0µ ±  7%  -29.62% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_1_step                             776.5µ ±  5%   540.7µ ±  1%  -30.37% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_100_steps                         1342.2µ ±  1%   980.7µ ±  1%  -26.94% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_1000_steps                         6.174m ±  2%   4.774m ±  1%  -22.68% (p=0.002 n=6)
Query/rate(a_100[1d]),_instant_query                                     43.27m ±  0%   33.47m ±  1%  -22.64% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_1_step                           43.87m ±  3%   33.98m ±  1%  -22.56% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_100_steps                        97.46m ±  2%   75.80m ±  8%  -22.22% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_1000_steps                       582.3m ±  1%   456.3m ±  2%  -21.64% (p=0.002 n=6)
Query/rate(a_2000[1d]),_instant_query                                    807.7m ±  1%   618.6m ±  1%  -23.41% (p=0.002 n=6)
Query/rate(a_2000[1d]),_range_query_with_1_step                          819.2m ±  2%   626.1m ±  0%  -23.58% (p=0.002 n=6)
Query/rate(a_2000[1d]),_range_query_with_100_steps                        1.912 ±  1%    1.477 ±  0%  -22.77% (p=0.002 n=6)
Query/rate(a_2000[1d]),_range_query_with_1000_steps                      11.606 ±  1%    9.155 ±  2%  -21.12% (p=0.002 n=6)
Query/sum(a_1),_instant_query                                            124.2µ ± 19%   115.6µ ±  3%   -6.89% (p=0.002 n=6)
Query/sum(a_1),_range_query_with_1_step                                  127.2µ ±  3%   115.7µ ±  1%   -9.04% (p=0.002 n=6)
Query/sum(a_1),_range_query_with_100_steps                               159.4µ ±  3%   123.4µ ±  1%  -22.59% (p=0.002 n=6)
Query/sum(a_1),_range_query_with_1000_steps                              454.9µ ±  3%   182.2µ ± 11%  -59.95% (p=0.002 n=6)
Query/sum(a_100),_instant_query                                          595.1µ ±  1%   569.2µ ±  3%   -4.36% (p=0.002 n=6)
Query/sum(a_100),_range_query_with_1_step                                605.2µ ±  2%   570.8µ ±  1%   -5.69% (p=0.002 n=6)
Query/sum(a_100),_range_query_with_100_steps                             1.225m ±  1%   1.027m ±  1%  -16.17% (p=0.002 n=6)
Query/sum(a_100),_range_query_with_1000_steps                            6.825m ±  1%   5.231m ±  0%  -23.35% (p=0.002 n=6)
Query/sum(a_2000),_instant_query                                         7.422m ±  2%   7.149m ±  1%   -3.68% (p=0.002 n=6)
Query/sum(a_2000),_range_query_with_1_step                               7.644m ±  1%   7.199m ±  1%   -5.82% (p=0.002 n=6)
Query/sum(a_2000),_range_query_with_100_steps                            18.88m ±  1%   15.82m ±  1%  -16.21% (p=0.002 n=6)
Query/sum(a_2000),_range_query_with_1000_steps                          154.65m ± 11%   92.52m ±  9%  -40.17% (p=0.002 n=6)
Query/sum_by_(l)(h_1),_instant_query                                     161.2µ ± 22%   153.6µ ± 11%        ~ (p=0.132 n=6)
Query/sum_by_(l)(h_1),_range_query_with_1_step                           159.5µ ±  8%   146.2µ ± 15%   -8.36% (p=0.026 n=6)
Query/sum_by_(l)(h_1),_range_query_with_100_steps                        233.6µ ±  2%   177.5µ ±  4%  -24.00% (p=0.002 n=6)
Query/sum_by_(l)(h_1),_range_query_with_1000_steps                       879.4µ ±  1%   486.3µ ±  3%  -44.70% (p=0.002 n=6)
Query/sum_by_(l)(h_100),_instant_query                                   2.652m ±  0%   2.536m ±  4%   -4.37% (p=0.002 n=6)
Query/sum_by_(l)(h_100),_range_query_with_1_step                         2.733m ±  1%   2.589m ±  1%   -5.28% (p=0.002 n=6)
Query/sum_by_(l)(h_100),_range_query_with_100_steps                      9.096m ±  1%   5.140m ±  1%  -43.49% (p=0.002 n=6)
Query/sum_by_(l)(h_100),_range_query_with_1000_steps                     67.28m ±  1%   29.10m ±  1%  -56.75% (p=0.002 n=6)
Query/sum_by_(l)(h_2000),_instant_query                                  47.19m ±  1%   46.02m ±  3%   -2.46% (p=0.041 n=6)
Query/sum_by_(l)(h_2000),_range_query_with_1_step                        49.58m ±  0%   46.65m ±  1%   -5.92% (p=0.002 n=6)
Query/sum_by_(l)(h_2000),_range_query_with_100_steps                    187.43m ±  1%   94.84m ±  1%  -49.40% (p=0.002 n=6)
Query/sum_by_(l)(h_2000),_range_query_with_1000_steps                   1565.5m ±  2%   557.1m ±  4%  -64.42% (p=0.002 n=6)
Query/sum_by_(le)(h_1),_instant_query                                    162.4µ ±  5%   151.5µ ±  4%   -6.73% (p=0.002 n=6)
Query/sum_by_(le)(h_1),_range_query_with_1_step                          171.7µ ±  8%   149.9µ ± 21%  -12.66% (p=0.041 n=6)
Query/sum_by_(le)(h_1),_range_query_with_100_steps                       362.6µ ±  6%   192.2µ ±  6%  -47.00% (p=0.002 n=6)
Query/sum_by_(le)(h_1),_range_query_with_1000_steps                     1950.2µ ±  3%   540.0µ ± 15%  -72.31% (p=0.002 n=6)
Query/sum_by_(le)(h_100),_instant_query                                  2.618m ±  3%   2.499m ±  9%        ~ (p=0.065 n=6)
Query/sum_by_(le)(h_100),_range_query_with_1_step                        2.745m ±  9%   2.520m ±  2%   -8.22% (p=0.002 n=6)
Query/sum_by_(le)(h_100),_range_query_with_100_steps                     6.919m ± 21%   5.337m ±  1%  -22.86% (p=0.002 n=6)
Query/sum_by_(le)(h_100),_range_query_with_1000_steps                    43.58m ± 11%   30.59m ±  7%  -29.82% (p=0.002 n=6)
Query/sum_by_(le)(h_2000),_instant_query                                 46.85m ±  1%   45.35m ±  5%        ~ (p=0.240 n=6)
Query/sum_by_(le)(h_2000),_range_query_with_1_step                       48.35m ± 11%   44.88m ±  8%   -7.19% (p=0.026 n=6)
Query/sum_by_(le)(h_2000),_range_query_with_100_steps                   126.73m ±  7%   97.80m ±  6%  -22.83% (p=0.002 n=6)
Query/sum_by_(le)(h_2000),_range_query_with_1000_steps                  1018.1m ± 10%   559.1m ±  0%  -45.09% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_instant_query                          161.9µ ±  2%   148.0µ ±  1%   -8.58% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_range_query_with_1_step                165.7µ ±  6%   159.8µ ±  4%   -3.56% (p=0.004 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_range_query_with_100_steps             345.5µ ±  2%   193.4µ ±  3%  -44.04% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_range_query_with_1000_steps           1817.7µ ±  1%   552.1µ ±  9%  -69.63% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_100[1m])),_instant_query                        2.518m ±  3%   2.387m ±  1%   -5.20% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_100[1m])),_range_query_with_1_step              2.649m ±  3%   2.439m ±  5%   -7.92% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_100[1m])),_range_query_with_100_steps           9.422m ±  4%   5.921m ±  2%  -37.15% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_100[1m])),_range_query_with_1000_steps          70.57m ±  1%   33.97m ±  0%  -51.86% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_2000[1m])),_instant_query                       44.42m ±  2%   42.93m ±  1%   -3.35% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_2000[1m])),_range_query_with_1_step             46.59m ±  1%   43.10m ±  1%   -7.50% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_2000[1m])),_range_query_with_100_steps          189.8m ±  1%   112.5m ±  1%  -40.72% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_2000[1m])),_range_query_with_1000_steps        1609.8m ±  0%   655.5m ±  1%  -59.28% (p=0.002 n=6)
geomean                                                                  5.722m         4.336m        -24.22%

                                                                 │      standard      │               streaming                │
                                                                 │        B/op        │      B/op       vs base                │
Query/a_1,_instant_query                                                20.24Ki ±  0%    17.25Ki ±  0%   -14.76% (p=0.002 n=6)
Query/a_1,_range_query_with_1_step                                      20.54Ki ±  0%    17.26Ki ±  0%   -15.98% (p=0.002 n=6)
Query/a_1,_range_query_with_100_steps                                   21.21Ki ±  1%    17.96Ki ±  0%   -15.28% (p=0.002 n=6)
Query/a_1,_range_query_with_1000_steps                                  27.36Ki ±  0%    25.87Ki ±  2%    -5.46% (p=0.002 n=6)
Query/a_100,_instant_query                                              140.7Ki ±  0%    122.3Ki ±  0%   -13.11% (p=0.002 n=6)
Query/a_100,_range_query_with_1_step                                    137.2Ki ±  0%    123.1Ki ±  0%   -10.22% (p=0.002 n=6)
Query/a_100,_range_query_with_100_steps                                 196.8Ki ±  0%    182.5Ki ±  0%    -7.29% (p=0.002 n=6)
Query/a_100,_range_query_with_1000_steps                                773.4Ki ±  0%    761.7Ki ±  1%    -1.52% (p=0.002 n=6)
Query/a_2000,_instant_query                                             2.423Mi ±  0%    2.336Mi ±  1%    -3.62% (p=0.002 n=6)
Query/a_2000,_range_query_with_1_step                                   2.344Mi ±  0%    2.514Mi ±  1%    +7.24% (p=0.002 n=6)
Query/a_2000,_range_query_with_100_steps                                3.434Mi ±  0%    3.350Mi ±  1%    -2.45% (p=0.002 n=6)
Query/a_2000,_range_query_with_1000_steps                               14.51Mi ±  0%    14.44Mi ±  1%    -0.48% (p=0.041 n=6)
Query/rate(a_1[1m]),_instant_query                                      22.42Ki ±  0%    17.65Ki ±  0%   -21.25% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_1_step                            22.85Ki ±  0%    17.66Ki ±  0%   -22.70% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_100_steps                         23.40Ki ±  0%    18.39Ki ±  0%   -21.42% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_1000_steps                        29.30Ki ±  1%    26.14Ki ±  2%   -10.77% (p=0.002 n=6)
Query/rate(a_100[1m]),_instant_query                                    165.2Ki ±  0%    125.2Ki ±  0%   -24.22% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_1_step                          161.6Ki ±  0%    126.1Ki ±  0%   -21.99% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_100_steps                       221.6Ki ±  0%    185.6Ki ±  0%   -16.21% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_1000_steps                      762.4Ki ±  0%    724.8Ki ±  1%    -4.93% (p=0.002 n=6)
Query/rate(a_2000[1m]),_instant_query                                   2.871Mi ±  0%    2.403Mi ±  1%   -16.28% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_1_step                         2.793Mi ±  0%    2.567Mi ±  3%    -8.08% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_100_steps                      3.878Mi ±  0%    3.408Mi ±  2%   -12.14% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_1000_steps                     14.33Mi ±  0%    13.91Mi ±  1%    -2.88% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_10000_steps                       84.54Ki ± 14%    91.79Ki ± 14%    +8.58% (p=0.004 n=6)
Query/rate(a_100[1m]),_range_query_with_10000_steps                     5.191Mi ±  1%    5.038Mi ±  5%         ~ (p=0.310 n=6)
Query/rate(a_2000[1m]),_range_query_with_10000_steps                    1.043Gi ±  0%    3.088Gi ±  0%  +195.99% (p=0.002 n=6)
Query/rate(a_1[1d]),_instant_query                                     665.90Ki ±  3%    63.47Ki ±  2%   -90.47% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_1_step                           666.50Ki ±  2%    63.51Ki ±  1%   -90.47% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_100_steps                        688.60Ki ±  5%    63.30Ki ±  2%   -90.81% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_1000_steps                       681.83Ki ±  4%    68.42Ki ±  5%   -89.97% (p=0.002 n=6)
Query/rate(a_100[1d]),_instant_query                                    5.204Mi ±  2%    4.475Mi ±  1%   -14.01% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_1_step                          5.220Mi ±  1%    4.504Mi ±  0%   -13.71% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_100_steps                       5.293Mi ±  3%    4.500Mi ±  1%   -14.98% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_1000_steps                      5.756Mi ±  5%    4.801Mi ±  6%   -16.60% (p=0.002 n=6)
Query/rate(a_2000[1d]),_instant_query                                   87.73Mi ±  1%    86.23Mi ±  1%    -1.71% (p=0.004 n=6)
Query/rate(a_2000[1d]),_range_query_with_1_step                         87.64Mi ±  1%    87.09Mi ±  1%    -0.63% (p=0.026 n=6)
Query/rate(a_2000[1d]),_range_query_with_100_steps                      92.23Mi ±  1%   119.18Mi ±  1%   +29.23% (p=0.002 n=6)
Query/rate(a_2000[1d]),_range_query_with_1000_steps                     126.9Mi ±  1%    406.9Mi ±  0%  +220.58% (p=0.002 n=6)
Query/sum(a_1),_instant_query                                           22.00Ki ±  0%    17.86Ki ±  0%   -18.81% (p=0.002 n=6)
Query/sum(a_1),_range_query_with_1_step                                 22.51Ki ±  0%    17.99Ki ±  0%   -20.11% (p=0.002 n=6)
Query/sum(a_1),_range_query_with_100_steps                              39.48Ki ±  0%    19.70Ki ±  0%   -50.10% (p=0.002 n=6)
Query/sum(a_1),_range_query_with_1000_steps                            195.02Ki ±  0%    35.83Ki ±  1%   -81.63% (p=0.002 n=6)
Query/sum(a_100),_instant_query                                         170.8Ki ±  0%    123.9Ki ±  0%   -27.47% (p=0.002 n=6)
Query/sum(a_100),_range_query_with_1_step                               170.9Ki ±  0%    124.2Ki ±  0%   -27.33% (p=0.002 n=6)
Query/sum(a_100),_range_query_with_100_steps                            247.7Ki ±  0%    185.8Ki ±  0%   -25.00% (p=0.002 n=6)
Query/sum(a_100),_range_query_with_1000_steps                           972.8Ki ±  0%    818.6Ki ±  1%   -15.85% (p=0.002 n=6)
Query/sum(a_2000),_instant_query                                        3.027Mi ±  0%    2.167Mi ±  1%   -28.42% (p=0.002 n=6)
Query/sum(a_2000),_range_query_with_1_step                              3.069Mi ±  0%    2.168Mi ±  0%   -29.34% (p=0.002 n=6)
Query/sum(a_2000),_range_query_with_100_steps                           4.174Mi ±  0%    3.287Mi ±  0%   -21.25% (p=0.002 n=6)
Query/sum(a_2000),_range_query_with_1000_steps                          15.46Mi ±  0%    14.85Mi ±  1%    -3.91% (p=0.002 n=6)
Query/sum_by_(l)(h_1),_instant_query                                    28.97Ki ±  0%    23.28Ki ±  0%   -19.64% (p=0.002 n=6)
Query/sum_by_(l)(h_1),_range_query_with_1_step                          29.43Ki ±  0%    23.39Ki ±  0%   -20.52% (p=0.002 n=6)
…